refactor base video parser
