+++ src/classes/index.ts
@@ -1,16 +1,16 @@
-export { default as Client, Client as ClientTypes } from "./Client";
-export { default as Base, BaseAttributes } from "./Base";
-export { default as BaseVideo, BaseVideoAttributes } from "./BaseVideo";
-export { default as VideoCompact } from "./VideoCompact";
-export { default as Thumbnails } from "./Thumbnails";
+export { BaseAttributes, default as Base } from "./Base";
+export * from "./BaseVideo";
 export { default as Channel } from "./Channel";
 export { default as ChannelCompact } from "./ChannelCompact";
-export { default as Comment } from "./Comment";
-export { default as Reply } from "./Reply";
 export { default as Chat } from "./Chat";
-export { default as PlaylistCompact } from "./PlaylistCompact";
-export { default as Playlist } from "./Playlist";
-export { default as Video } from "./Video";
+export { Client as ClientTypes, default as Client } from "./Client";
+export { default as Comment } from "./Comment";
 export { default as LiveVideo } from "./LiveVideo";
-export { default as SearchResult } from "./SearchResult";
 export { default as MixPlaylist } from "./MixPlaylist";
+export { default as Playlist } from "./Playlist";
+export { default as PlaylistCompact } from "./PlaylistCompact";
+export { default as Reply } from "./Reply";
+export { default as SearchResult } from "./SearchResult";
+export { default as Thumbnails } from "./Thumbnails";
+export { default as Video } from "./Video";
+export { default as VideoCompact } from "./VideoCompact";
